# Array container alloc optimizations