@jacksonrnewhouse 2020-10-21
Agreed, it can be capped. I made the change and got the following benchmarks, showing it reduces the size of allocations without any impact on performance. This compares the PR I initially put up with the current state.
```
benchstat experiment.txt new_experiment.txt
name                                   old time/op    new time/op    delta
EvenIntervalArrayUnions-16               15.7µs ± 5%    15.9µs ± 4%     ~     (p=0.421 n=5+5)
InPlaceArrayUnions-16                     300µs ± 2%     301µs ± 3%     ~     (p=0.690 n=5+5)
AntagonisticArrayUnionsGrowth-16         29.2µs ± 4%    27.8µs ± 4%   -4.88%  (p=0.008 n=5+5)
RepeatedGrowthArrayUnion-16              3.49ms ± 7%    3.49ms ± 2%     ~     (p=0.548 n=5+5)
RepeatedSelfArrayUnion-16                6.42ms ± 1%    6.55ms ± 2%     ~     (p=0.222 n=5+5)
ArrayUnionThreshold/mostly-overlap-16    7.88µs ± 3%    7.67µs ± 2%   -2.74%  (p=0.016 n=5+5)
ArrayUnionThreshold/little-overlap-16    9.61µs ± 1%    9.45µs ± 4%     ~     (p=0.151 n=5+5)
ArrayUnionThreshold/no-overlap-16        18.3µs ± 2%    18.6µs ± 4%     ~     (p=0.690 n=5+5)

name                                   old alloc/op   new alloc/op   delta
EvenIntervalArrayUnions-16               2.96kB ± 0%    2.96kB ± 0%     ~     (all equal)
InPlaceArrayUnions-16                    15.6kB ± 0%    15.6kB ± 0%   -0.02%  (p=0.008 n=5+5)
AntagonisticArrayUnionsGrowth-16         49.3kB ± 0%    33.0kB ± 0%  -33.20%  (p=0.008 n=5+5)
RepeatedGrowthArrayUnion-16              48.0kB ± 0%    48.0kB ± 0%     ~     (all equal)
RepeatedSelfArrayUnion-16                41.0kB ± 0%    24.6kB ± 0%  -39.94%  (p=0.008 n=5+5)
ArrayUnionThreshold/mostly-overlap-16    26.0kB ± 0%    21.9kB ± 0%  -15.74%  (p=0.008 n=5+5)
ArrayUnionThreshold/little-overlap-16    26.0kB ± 0%    21.9kB ± 0%  -15.74%  (p=0.008 n=5+5)
ArrayUnionThreshold/no-overlap-16        34.2kB ± 0%    30.1kB ± 0%  -11.96%  (p=0.008 n=5+5)

name                                   old allocs/op  new allocs/op  delta
EvenIntervalArrayUnions-16                 7.00 ± 0%      7.00 ± 0%     ~     (all equal)
InPlaceArrayUnions-16                      9.00 ± 0%      9.00 ± 0%     ~     (all equal)
AntagonisticArrayUnionsGrowth-16           9.00 ± 0%      9.00 ± 0%     ~     (all equal)
RepeatedGrowthArrayUnion-16                32.0 ± 0%      32.0 ± 0%     ~     (all equal)
RepeatedSelfArrayUnion-16                  6.00 ± 0%      6.00 ± 0%     ~     (all equal)
ArrayUnionThreshold/mostly-overlap-16      7.00 ± 0%      7.00 ± 0%     ~     (all equal)
ArrayUnionThreshold/little-overlap-16      7.00 ± 0%      7.00 ± 0%     ~     (all equal)
ArrayUnionThreshold/no-overlap-16          9.00 ± 0%      9.00 ± 0%     ~     (all equal)
```