@lemire 2020-10-21
Let us merge and release.